+++ docs/pages/map-providers/_meta.js
@@ -5,5 +5,6 @@ export default {
   tms: "TMS",
   wms: "WMS",
   oam: "OpenAerialMap",
+  google: "Google Maps",
   "serving-raster-tiles": "Serving Raster Tiles",
 };